feat: Refactor charm to use WorkloadBuilder
Given Juju's distributed system characteristics, using the builder
pattern for determining if a workload's preconditions are met has
several advantages. The primary advantages are that we can ask the
builder if it is ready to proceed to workload start-up and it
simplifies debugging because we can ask the builder to share its
internal state.

This change also moves away from a sepearte python file for the app
and the charm because it was too noisy. In time we may move back but
that will need to be reassed.

Finally, as part of this change the charm now supports a database
relation.