feat: Separate workload configuration and charm code
1. Use a WorkloadAgentBuilder to collect values as events are recieved to
simplify checking when we are ready to configure and start the charm.
Additionally, this means that most code inside charm.py is now focused
on the charm event management and code specific to the workload is
contained in workload.py.

2. Add a simple grafana_dashboard example.

3. Add ingress configuration via the traefik-route interface. This is
essentially a workaround for the traefik charm not exposing richer
configuration options.